Implemented a couple convenience functions for testing.
Note: The function gamma_corr previously only filled the upper half of the matrix it returned. This was OK, because the methods using the function used the symmetry of the matrix to retrieve values in the lower half. The function now fills the entire matrix, because only filling the upper half is just begging for bugs to appear in the future.