QuantumMie now passes test suite.
Have overridden the multithreading methods to ensure that temperature is only set once for all collision integrals, and use a private data member current_temperature to keep track of what temperature the current computation is running for.

The method set_temperature is the ONE AND ONLY method that should be used to set the model temperature, and all public methods should call this method immediately when they are invoked. This way, the class appears stateless to an end-user. Note, however, that it is NOT thread-safe. Running computations using the QuantumMie class on several threads with different temperatures will lead to nonsensical results.

Also fixed bug in Sutherland resulting from vdw_alpha not being reset properly when the coefficients were changed and it was recomputed.