Modify parameter handling on input and add set-methods.
Init now ensures that the equation of state uses the same parameters as the RET model (if applicable). Made some modifications to get_Eij, which ensures that the output is the same for a single component mixture, and a binary mixture of identical components, also if x != [0.5, 0.5], which is the expected behaviour. Previously, pure fluid calculation in get_Eij used a hardcoded value of x=[0.5, 0.5], which could lead to confusing output when testing the method.