Implemented all considered transfer length models
As of this commit, the different transfer lengths are accessed by using cpp_kingas.set_cd_model. The "old" model is model 0, the new MTL is model 2, the ETL is model 3. After this commit, the structure will be cleaned up so that the MTL, ETL and collision diameter are separate methods.