Finished first implementation of Axilrod-Teller correction for Tang-Toennies potential
Ensured that AxilrodTellerCorrected classes properly override the set_internals method of KineticGas,
as well as the get_omegapoint and get_transfer_length_point.

Ensured that the potential(int, int, double) methods are properly overriden and hidden so that density dependence
is ensured through calls to set_internals.

Did some tweaking on the set/get methods for transfer length models for better handling of invalid models.