Isolate thread-safe caching logic in dedicated KineticGasCache class.
This allows us to make a shitload of stuff "const" that should have been made const years ago. Most of the changes here are just making a bunch of stuff from the collision integrals (omega) downwards const-qualified.

Note that the cache is "mutable", so values can be stored in the cache even if they're used in the collision integrals or transfer lengths.

The whole point here is that omega, which is multithreaded, is const-qualified, so if you want to cache anything below "omega", you need to use the dedicated (thread-safe) caching mechanism, which is mutable.